Add support for Redis-Sentinel
To aid things such as specifying passwords or databases for the sentinels, a new RedisSetting struct has been created.
This can hold information such as Host or Database so that parsing a URI is not a requirement.

The URI option has been left in on the WorkerSettings for backwards compatibility

Tests have been updated so that we can individually test each connection